#337 Changes functionality behind InteractionDetector to show that it is not just for the player, also allowed it to filter out entities if need be. Moved it out of the player component folder.
