+++ source/core/src/main/com/csse3200/game/areas/SpaceGameArea.java
@@ -227,7 +227,7 @@ public class SpaceGameArea extends GameArea {
 
     player = spawnPlayer();
     player.getComponent(PlayerActions.class).setGameMap(gameMap);
-    player.getComponent(InventoryComponent.class).addItem(ItemFactory.createGateItem());
+    player.getComponent(InventoryComponent.class).addItem(ItemFactory.createCowFood());
 
     tractor = spawnTractor();
     spawnQuestgiver();
